pyproject.toml
[project]
name = "auto-translation-tools"
version = "0.1.0"
description = "自动化翻译工具平台：专名提取翻译、Gemini API调用等可扩展工具集"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Aldo", email = "aldo@example.com"}
]
keywords = ["translation", "nlp", "ner", "gemini", "corenlp"]

dependencies = [
    "httpx>=0.27.0",
    "google-genai>=1.0.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.1.0",
]

[project.scripts]
name-extractor = "auto_translation_tools.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/auto_translation_tools"]

[tool.ruff]
line-length = 100
target-version = "py310"
